made review component
